+++ components/04-views/dashboard/gift-card/gift-card.config.js
@@ -7,6 +7,8 @@ module.exports = {
           input: {
             class: 'dashboard-form__divider',
             field: {
+              id: 'giftCardCode',
+              name: 'giftCardCode',
               placeholder: 'Enter gitf card code'
             }
           }